Add files via upload
A script that develops and validates a linear mixed effects model that can be applied to satellite databases to predict ocean carbon gels. NOTE: this model was produced in R, while the rest of this repository is coded in Python. R has superior capabilities for developing linear mixed effects models.