Add files via upload
Added code that inputs python data frames into R. In this case, it was a large matrix of satellite data. I did this in order to make the predictions (of a linear mixed effect model) in R, as opposed to python. R is superior in terms of statistical predictions and analysis. However, I prefer to plot the predictions in python.